Extract _artifacts.py from _legacy.py
